# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server of University project",
        "title": "University app",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dias Utekin",
            "email": "dias.utekin@nu.edu.kz"
        },
        "version": "1.0"
    },
    "host": "ec2-18-192-183-17.eu-central-1.compute.amazonaws.com:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/students/create": {
            "post": {
                "description": "Create Student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "CreateStudent",
                "operationId": "create-student",
                "parameters": [
                    {
                        "description": "create account",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Studio56School_university_internal_model.Student"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Studio56School_university_internal_model.Student"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Studio56School_university_internal_model.Student": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string",
                    "example": "F"
                },
                "name": {
                    "type": "string",
                    "example": "Madina"
                },
                "surname": {
                    "type": "string",
                    "example": "Akhmetova"
                }
            }
        }
    }
}